Make overwritting of access filters to actions optional.
In some cases it is convenient to combine access filter to
a certain action of a controller. This was not possible
since an access filter for the same action would overwrite
any access filter to this action which had been defined earlier.

Make this behavior optional but default to overwitting.